Improvement: Improve code quality splitting methods in Megabot
Split methods in Megabot into Gui/GuiCommandHandler.java
