+++ data/duke.txt
@@ -1,3 +1,4 @@
 T | 0 | read book
 D | 0 | return book | 2023-12-01 0000
 T | 0 | task
+T | 0 | meow
